Create RSpec tutorial project!
Setup the tutorial project shell:

  - Add general .gitignore
  - Specify the Ruby version as 2.1.0
  - Add a Gemfile for future use
  - Include GPLv3 license

A goal of this project is to give back to the RSpec (and Ruby) community
as a whole. We believe in fostering a welcoming and positive community
experience. Due to these beliefs we are pledging our respect and
appreciation for contributions made to this project. Our pledge is
included in the CODE_OF_CONDUCT.md file in the project root.

For those wishing to contribute, a more detailed guide on our
contribution process is outlined in CONTRIBUTING.md.